// src/generated/proto/stream_list.ts
export interface LiveChatMessageListRequest {
  liveChatId?: string | undefined;
  hl?: string | undefined;
  profileImageSize?: number | undefined;
  maxResults?: number | undefined;
  pageToken?: string | undefined;
  part: string[];
}

export interface LiveChatMessageListResponse {
  kind?: string | undefined;
  etag?: string | undefined;
  offlineAt?: string | undefined;
  pageInfo?: PageInfo | undefined;
  nextPageToken?: string | undefined;
  items: LiveChatMessage[];
  activePollItem?: LiveChatMessage | undefined;
}

export interface LiveChatMessage {
  kind?: string | undefined;
  etag?: string | undefined;
  id?: string | undefined;
  snippet?: LiveChatMessageSnippet | undefined;
  authorDetails?: LiveChatMessageAuthorDetails | undefined;
}

export interface LiveChatMessageAuthorDetails {
  channelId?: string | undefined;
  channelUrl?: string | undefined;
  displayName?: string | undefined;
  profileImageUrl?: string | undefined;
  isVerified?: boolean | undefined;
  isChatOwner?: boolean | undefined;
  isChatSponsor?: boolean | undefined;
  isChatModerator?: boolean | undefined;
}

export interface LiveChatMessageSnippet {
  type?: LiveChatMessageSnippet_TypeWrapper_Type | undefined;
  liveChatId?: string | undefined;
  authorChannelId?: string | undefined;
  publishedAt?: string | undefined;
  hasDisplayContent?: boolean | undefined;
  displayMessage?: string | undefined;
  textMessageDetails?: LiveChatTextMessageDetails | undefined;
  messageDeletedDetails?: LiveChatMessageDeletedDetails | undefined;
  messageRetractedDetails?: LiveChatMessageRetractedDetails | undefined;
  userBannedDetails?: LiveChatUserBannedMessageDetails | undefined;
  superChatDetails?: LiveChatSuperChatDetails | undefined;
  superStickerDetails?: LiveChatSuperStickerDetails | undefined;
  newSponsorDetails?: LiveChatNewSponsorDetails | undefined;
  memberMilestoneChatDetails?: LiveChatMemberMilestoneChatDetails | undefined;
  membershipGiftingDetails?: LiveChatMembershipGiftingDetails | undefined;
  giftMembershipReceivedDetails?: LiveChatGiftMembershipReceivedDetails | undefined;
  pollDetails?: LiveChatPollDetails | undefined;
}

export interface LiveChatMessageSnippet_TypeWrapper {
}

export enum LiveChatMessageSnippet_TypeWrapper_Type {
  INVALID_TYPE = 0,
  TEXT_MESSAGE_EVENT = 1,
  TOMBSTONE = 2,
  FAN_FUNDING_EVENT = 3,
  CHAT_ENDED_EVENT = 4,
  SPONSOR_ONLY_MODE_STARTED_EVENT = 5,
  SPONSOR_ONLY_MODE_ENDED_EVENT = 6,
  NEW_SPONSOR_EVENT = 7,
  MEMBER_MILESTONE_CHAT_EVENT = 17,
  MEMBERSHIP_GIFTING_EVENT = 18,
  GIFT_MEMBERSHIP_RECEIVED_EVENT = 19,
  MESSAGE_DELETED_EVENT = 8,
  MESSAGE_RETRACTED_EVENT = 9,
  USER_BANNED_EVENT = 10,
  SUPER_CHAT_EVENT = 15,
  SUPER_STICKER_EVENT = 16,
  POLL_EVENT = 20,
  UNRECOGNIZED = -1,
}
// ...
export const LiveChatMessageListRequest: MessageFns<LiveChatMessageListRequest> = {
  encode(message: LiveChatMessageListRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.liveChatId !== undefined && message.liveChatId !== "") {
      writer.uint32(10).string(message.liveChatId);
    }
    if (message.hl !== undefined && message.hl !== "") {
      writer.uint32(18).string(message.hl);
    }
    if (message.profileImageSize !== undefined && message.profileImageSize !== 0) {
      writer.uint32(24).uint32(message.profileImageSize);
    }
    if (message.maxResults !== undefined && message.maxResults !== 0) {
      writer.uint32(784).uint32(message.maxResults);
    }
    if (message.pageToken !== undefined && message.pageToken !== "") {
      writer.uint32(794).string(message.pageToken);
    }
    for (const v of message.part) {
      writer.uint32(802).string(v!);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): LiveChatMessageListRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseLiveChatMessageListRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.liveChatId = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.hl = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.profileImageSize = reader.uint32();
          continue;
        }
        case 98: {
          if (tag !== 784) {
            break;
          }

          message.maxResults = reader.uint32();
          continue;
        }
        case 99: {
          if (tag !== 794) {
            break;
          }

          message.pageToken = reader.string();
          continue;
        }
        case 100: {
          if (tag !== 802) {
            break;
          }

          message.part.push(reader.string());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): LiveChatMessageListRequest {
    return {
      liveChatId: isSet(object.liveChatId) ? globalThis.String(object.liveChatId) : "",
      hl: isSet(object.hl) ? globalThis.String(object.hl) : "",
      profileImageSize: isSet(object.profileImageSize) ? globalThis.Number(object.profileImageSize) : 0,
      maxResults: isSet(object.maxResults) ? globalThis.Number(object.maxResults) : 0,
      pageToken: isSet(object.pageToken) ? globalThis.String(object.pageToken) : "",
      part: globalThis.Array.isArray(object?.part) ? object.part.map((e: any) => globalThis.String(e)) : [],
    };
  },

  toJSON(message: LiveChatMessageListRequest): unknown {
    const obj: any = {};
    if (message.liveChatId !== undefined && message.liveChatId !== "") {
      obj.liveChatId = message.liveChatId;
    }
    if (message.hl !== undefined && message.hl !== "") {
      obj.hl = message.hl;
    }
    if (message.profileImageSize !== undefined && message.profileImageSize !== 0) {
      obj.profileImageSize = Math.round(message.profileImageSize);
    }
    if (message.maxResults !== undefined && message.maxResults !== 0) {
      obj.maxResults = Math.round(message.maxResults);
    }
    if (message.pageToken !== undefined && message.pageToken !== "") {
      obj.pageToken = message.pageToken;
    }
    if (message.part?.length) {
      obj.part = message.part;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<LiveChatMessageListRequest>, I>>(base?: I): LiveChatMessageListRequest {
    return LiveChatMessageListRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<LiveChatMessageListRequest>, I>>(object: I): LiveChatMessageListRequest {
    const message = createBaseLiveChatMessageListRequest();
    message.liveChatId = object.liveChatId ?? "";
    message.hl = object.hl ?? "";
    message.profileImageSize = object.profileImageSize ?? 0;
    message.maxResults = object.maxResults ?? 0;
    message.pageToken = object.pageToken ?? "";
    message.part = object.part?.map((e) => e) || [];
    return message;
  },
};
// ...
export const LiveChatMessage: MessageFns<LiveChatMessage> = {
  encode(message: LiveChatMessage, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.kind !== undefined && message.kind !== "") {
      writer.uint32(1602).string(message.kind);
    }
    if (message.etag !== undefined && message.etag !== "") {
      writer.uint32(1610).string(message.etag);
    }
    if (message.id !== undefined && message.id !== "") {
      writer.uint32(810).string(message.id);
    }
    if (message.snippet !== undefined) {
      LiveChatMessageSnippet.encode(message.snippet, writer.uint32(18).fork()).join();
    }
    if (message.authorDetails !== undefined) {
      LiveChatMessageAuthorDetails.encode(message.authorDetails, writer.uint32(26).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): LiveChatMessage {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseLiveChatMessage();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 200: {
          if (tag !== 1602) {
            break;
          }

          message.kind = reader.string();
          continue;
        }
        case 201: {
          if (tag !== 1610) {
            break;
          }

          message.etag = reader.string();
          continue;
        }
        case 101: {
          if (tag !== 810) {
            break;
          }

          message.id = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.snippet = LiveChatMessageSnippet.decode(reader, reader.uint32());
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.authorDetails = LiveChatMessageAuthorDetails.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): LiveChatMessage {
    return {
      kind: isSet(object.kind) ? globalThis.String(object.kind) : "",
      etag: isSet(object.etag) ? globalThis.String(object.etag) : "",
      id: isSet(object.id) ? globalThis.String(object.id) : "",
      snippet: isSet(object.snippet) ? LiveChatMessageSnippet.fromJSON(object.snippet) : undefined,
      authorDetails: isSet(object.authorDetails)
        ? LiveChatMessageAuthorDetails.fromJSON(object.authorDetails)
        : undefined,
    };
  },

  toJSON(message: LiveChatMessage): unknown {
    const obj: any = {};
    if (message.kind !== undefined && message.kind !== "") {
      obj.kind = message.kind;
    }
    if (message.etag !== undefined && message.etag !== "") {
      obj.etag = message.etag;
    }
    if (message.id !== undefined && message.id !== "") {
      obj.id = message.id;
    }
    if (message.snippet !== undefined) {
      obj.snippet = LiveChatMessageSnippet.toJSON(message.snippet);
    }
    if (message.authorDetails !== undefined) {
      obj.authorDetails = LiveChatMessageAuthorDetails.toJSON(message.authorDetails);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<LiveChatMessage>, I>>(base?: I): LiveChatMessage {
    return LiveChatMessage.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<LiveChatMessage>, I>>(object: I): LiveChatMessage {
    const message = createBaseLiveChatMessage();
    message.kind = object.kind ?? "";
    message.etag = object.etag ?? "";
    message.id = object.id ?? "";
    message.snippet = (object.snippet !== undefined && object.snippet !== null)
      ? LiveChatMessageSnippet.fromPartial(object.snippet)
      : undefined;
    message.authorDetails = (object.authorDetails !== undefined && object.authorDetails !== null)
      ? LiveChatMessageAuthorDetails.fromPartial(object.authorDetails)
      : undefined;
    return message;
  },
};
// ...
export interface V3DataLiveChatMessageServiceClient extends Client {
  /** Allows a user to load live chat through a server-streamed RPC. */
  streamList(
    request: LiveChatMessageListRequest,
    options?: Partial<CallOptions>,
  ): ClientReadableStream<LiveChatMessageListResponse>;
  streamList(
    request: LiveChatMessageListRequest,
    metadata?: Metadata,
    options?: Partial<CallOptions>,
  ): ClientReadableStream<LiveChatMessageListResponse>;
}

export const V3DataLiveChatMessageServiceClient = makeGenericClientConstructor(
  V3DataLiveChatMessageServiceService,
  "youtube.api.v3.V3DataLiveChatMessageService",
) as unknown as {
  new (
    address: string,
    credentials: ChannelCredentials,
    options?: Partial<ClientOptions>,
  ): V3DataLiveChatMessageServiceClient;
  service: typeof V3DataLiveChatMessageServiceService;
  serviceName: string;
};
// ...
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };